refactor libraries based on conditional directive changes
