+++ example/example.dart
@@ -2,31 +2,30 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-/**
- * A simple web app to hand-test the usage library.
- */
+/// A simple web app to hand-test the usage library.
 library usage_example;
 
+import 'dart:async';
 import 'dart:html';
 
-import 'package:usage/usage_html.dart';
+import 'package:usage/usage.dart';
 
 Analytics _analytics;
 String _lastUa;
 int _count = 0;
 
 void main() {
-  querySelector('#foo').onClick.listen((_) => _handleFoo(getAnalytics()));
-  querySelector('#bar').onClick.listen((_) => _handleBar(getAnalytics()));
-  querySelector('#page').onClick.listen((_) => _changePage(getAnalytics()));
+  querySelector('#foo').onClick.listen((_) => _handleFoo());
+  querySelector('#bar').onClick.listen((_) => _handleBar());
+  querySelector('#page').onClick.listen((_) => _changePage());
 }
 
 String _ua() => (querySelector('#ua') as InputElement).value.trim();
 
-Analytics getAnalytics() {
+Future<Analytics> getAnalytics() async {
   if (_analytics == null || _lastUa != _ua()) {
     _lastUa = _ua();
-    _analytics = new AnalyticsHtml(_lastUa, 'Test app', '1.0');
+    _analytics = await Analytics.create(_lastUa, 'Test app', '1.0');
     _analytics.optIn = true;
     _analytics.sendScreenView(window.location.pathname);
   }
@@ -34,15 +33,18 @@ Analytics getAnalytics() {
   return _analytics;
 }
 
-void _handleFoo(Analytics analytics) {
+Future _handleFoo() async {
+  Analytics analytics = await getAnalytics();
   analytics.sendEvent('main', 'foo');
 }
 
-void _handleBar(Analytics analytics) {
+Future _handleBar() async {
+  Analytics analytics = await getAnalytics();
   analytics.sendEvent('main', 'bar');
 }
 
-void _changePage(Analytics analytics) {
+Future _changePage() async {
+  Analytics analytics = await getAnalytics();
   window.history.pushState(null, 'new page', '${++_count}.html');
   analytics.sendScreenView(window.location.pathname);
 }
