+++ example/flutter_example/pubspec.yaml
@@ -4,4 +4,4 @@ dependencies:
   usage:
     path: ../..
   flutter:
-    path: ../../../../../Users/jackson/git/flutter/packages/flutter
+    path: ../../../flutter/packages/flutter
