+++ lib/src/uuid.dart
@@ -2,9 +2,7 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-/**
- * A UUID generator library.
- */
+/// A UUID generator library.
 library usage.uuid;
 
 import 'dart:math' show Random;
