+++ readme.md
@@ -1,23 +1,21 @@
 # usage
 
-`usage` is a wrapper around Google Analytics for both command-line apps and web
-apps.
+`usage` is a wrapper around Google Analytics for both command-line, web, and
+Flutter apps.
 
 [![Build Status](https://travis-ci.org/dart-lang/usage.svg)](https://travis-ci.org/dart-lang/usage)
 [![Coverage Status](https://img.shields.io/coveralls/dart-lang/usage.svg)](https://coveralls.io/r/dart-lang/usage?branch=master)
 
-## For web apps
+## Using this library
 
-To use this library as a web app, import the `usage_html.dart` library and
-instantiate the `AnalyticsHtml` class.
+In order to use this library, call the [Analytics.create] static method.
+You'll get either the command-line, web, or Flutter implementation based on
+the current platform.
 
 When you are creating a new property at [google analytics](https://www.google.com/analytics/)
 make sure to select not the website option, but the **mobile app** option.
 
 ## For command-line apps
-
-To use this library as a command-line app, import the `usage_io.dart` library
-and instantiate the `AnalyticsIO` class.
 
 Note, for CLI apps, the usage library will send analytics pings asynchronously.
 This is useful it that it doesn't block the app generally. It does have one
@@ -40,10 +38,10 @@ analytics.waitForLastPing(timeout: new Duration(milliseconds: 500)).then((_) {
 
 ## Using the API
 
-Import the package (in this example we use the `dart:io` version):
+Import the package:
 
 ```dart
-import 'package:usage/usage_io.dart';
+import 'package:usage/usage.dart';
 ```
 
 And call some analytics code:
@@ -51,7 +49,7 @@ And call some analytics code:
 ```dart
 final String UA = ...;
 
-Analytics ga = new AnalyticsIO(UA, 'ga_test', '1.0');
+Analytics ga = await Analytics.create(UA, 'ga_test', '1.0');
 ga.optIn = true;
 
 ga.sendScreenView('home');
